# frontend/src/assets/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Player Management","description":"Endpoints for creating and retrieving players"}],"paths":{"/api/players/{id}/stats":{"put":{"tags":["Player Management"],"summary":"Update player's stats","description":"Update a player's stats from a game round.","operationId":"updatePlayerStats","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlayerStats"}}},"required":true},"responses":{"201":{"description":"Update stats successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"Invalid player stats data provided","content":{"text/plain":{}}}}}},"/api/players/{id}/reset-stats":{"put":{"tags":["Player Management"],"summary":"Reset player's score","description":"Resets the wins and games played for a specific player to zero. Acting as a 'new' player for ranking purposes.","operationId":"resetPlayerStats","parameters":[{"name":"id","in":"path","description":"ID of the player to reset score for","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Player not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"200":{"description":"Player score reset successfully","content":{"application/json":{}}}}}},"/api/players":{"get":{"tags":["Player Management"],"summary":"Get all registered players","description":"Retrieves a list of all existing players in the game.","operationId":"getAllPlayers","responses":{"200":{"description":"Successfully retrieved list of players","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlayerSimplifiedDto"}}}}},"404":{"description":"Players not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlayerSimplifiedDto"}}}}}}},"post":{"tags":["Player Management"],"summary":"Create a new player","description":"Registers a new player in the game with a unique name and an optional icon.","operationId":"createPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"201":{"description":"Player created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"Invalid player data provided","content":{"text/plain":{}}}}}},"/api/players/{id}":{"get":{"tags":["Player Management"],"summary":"Get player by ID","description":"Retrieves a single player's details by their ID.","operationId":"getPlayerById","parameters":[{"name":"id","in":"path","description":"ID of the player to retrieve","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Player found","content":{"application/json":{}}},"404":{"description":"Player not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/api/players/leaderboard-stats":{"get":{"tags":["Player Management"],"summary":"Get Leaderboard","description":"Retrieves the Leaderboard by Wilson Score.","operationId":"getLeaderboardStats","responses":{"200":{"description":"Successfully retrieved leaderboard","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LeaderboardPlayerStatsDto"}}}}}}}}},"components":{"schemas":{"PlayerStats":{"required":["computerHistory","computerScore","playerHistory","playerScore","totalRounds"],"type":"object","properties":{"playerScore":{"minimum":0,"type":"integer","description":"The player's current score (points for wins)","format":"int32","example":1},"computerScore":{"minimum":0,"type":"integer","description":"The computer's current score (points from player's losses)","format":"int32","example":1},"playerHistory":{"type":"array","description":"History of player's choices (e.g., 'rock', 'paper')","example":["scissors","scissors"],"items":{"type":"string","description":"History of player's choices (e.g., 'rock', 'paper')","example":"[\"scissors\",\"scissors\"]","enum":["rock","paper","scissors","well"]}},"computerHistory":{"type":"array","description":"History of computer's choices against this player","example":["paper","rock"],"items":{"type":"string","description":"History of computer's choices against this player","example":"[\"paper\",\"rock\"]","enum":["rock","paper","scissors","well"]}},"totalRounds":{"minimum":0,"type":"integer","description":"Total number of rounds played by this player","format":"int32","example":2}},"description":"Detailed game statistics for a player"},"Player":{"required":["icon","name","stats"],"type":"object","properties":{"id":{"type":"integer","description":"Unique ID of the player","format":"int64","example":1},"name":{"type":"string","description":"The player's chosen name","example":"Eva"},"icon":{"type":"string","description":"An emoji or icon representing the player","example":"🧑🏻‍✈️"},"stats":{"$ref":"#/components/schemas/PlayerStats"}},"description":"Represents a player in the RPS game"},"PlayerSimplifiedDto":{"required":["icon","id","name"],"type":"object","properties":{"id":{"type":"integer","description":"Unique identifier of the player","format":"int64"},"name":{"type":"string","description":"Name of the player"},"icon":{"type":"string","description":"Icon representing the player"}},"description":"Summary of a player's basic information"},"LeaderboardPlayerStatsDto":{"required":["gamesPlayed","icon","id","name","score","winPercentage"],"type":"object","properties":{"id":{"type":"integer","description":"Unique identifier of the player","format":"int64","example":1},"name":{"type":"string","description":"Name of the player","example":"Woman Pilot"},"icon":{"type":"string","description":"Emoji icon representing the player","example":"✈️"},"winPercentage":{"type":"number","description":"Win percentage of the player","format":"double","example":77.0},"gamesPlayed":{"minimum":0,"type":"integer","description":"Total number of games played by the player","format":"int32","example":33},"score":{"type":"number","description":"Calculated score based on Wilson Score Interval, used for ranking","format":"double","example":0.2117}},"description":"Data Transfer Object for Leaderboard Player Stats displayed on the Leaderboard."}}}}